+++ sites/geohub/src/components/pages/map/layers/raster/RasterTransform.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
 	import RasterTransformAdvanced from '$components/pages/map/layers/raster/RasterTransformAdvanced.svelte';
 	import RasterTransformSimple from '$components/pages/map/layers/raster/RasterTransformSimple.svelte';
+	import { handleEnterKey } from '$lib/helper';
 	import type { Layer } from '$lib/types';
-	import { Tabs } from '@undp-data/svelte-undp-design';
 
 	export let layer: Layer;
 
@@ -36,7 +36,29 @@
 </script>
 
 <nav class="block">
-	<Tabs bind:tabs bind:activeTab fontSize="small" />
+	<div class="tabs is-centered is-toggle">
+		<ul>
+			{#each tabs as tab}
+				<li class={activeTab === tab.label ? 'is-active' : ''}>
+					<!-- svelte-ignore a11y-missing-attribute -->
+					<a
+						class="has-text-weight-bold"
+						role="tab"
+						tabindex="0"
+						data-sveltekit-preload-code="off"
+						data-sveltekit-preload-data="off"
+						on:click={() => {
+							activeTab = tab.label;
+						}}
+						on:keydown={handleEnterKey}
+					>
+						<span class="icon is-small"><i class={tab.icon} aria-hidden="true"></i></span>
+						<span class="is-capitalized">{tab.label}</span>
+					</a>
+				</li>
+			{/each}
+		</ul>
+	</div>
 
 	<div class="block" />
 	<p>