+++ sites/geohub/src/components/pages/map/layers/vector/VectorLabelPanel.svelte
@@ -1,29 +1,51 @@
 <script lang="ts">
-	import SymbolPlacement from '$components/maplibre/symbol/SymbolPlacement.svelte';
 	import TextColor from '$components/maplibre/symbol/TextColor.svelte';
 	import TextField from '$components/maplibre/symbol/TextField.svelte';
 	import TextFieldDecimalPosition from '$components/maplibre/symbol/TextFieldDecimalPosition.svelte';
-	import TextHaloCalor from '$components/maplibre/symbol/TextHaloColor.svelte';
-	import TextHaloWidth from '$components/maplibre/symbol/TextHaloWidth.svelte';
-	import TextMaxWidth from '$components/maplibre/symbol/TextMaxWidth.svelte';
+	import TextFont from '$components/maplibre/symbol/TextFont.svelte';
 	import TextSize from '$components/maplibre/symbol/TextSize.svelte';
-	import { getLayerStyle, getPropertyValueFromExpression, getTextFieldDataType } from '$lib/helper';
-	import type { Layer } from '$lib/types';
-	import { MAPSTORE_CONTEXT_KEY, type MapStore } from '$stores';
+	import VectorLabelPropertyEditor from '$components/maplibre/vector/VectorLabelPropertyEditor.svelte';
+	import FieldControl from '$components/util/FieldControl.svelte';
+	import {
+		getLayerStyle,
+		getPropertyValueFromExpression,
+		getTextFieldDataType,
+		handleEnterKey
+	} from '$lib/helper';
+	import type { Layer, VectorTileMetadata } from '$lib/types';
+	import {
+		COLORMAP_NAME_CONTEXT_KEY_LABEL,
+		DEFAULTCOLOR_CONTEXT_KEY_LABEL,
+		MAPSTORE_CONTEXT_KEY,
+		NUMBER_OF_CLASSES_CONTEXT_KEY_LABEL,
+		type MapStore
+	} from '$stores';
 	import type { LayerSpecification } from 'maplibre-gl';
 	import { getContext, onMount } from 'svelte';
 
 	const map: MapStore = getContext(MAPSTORE_CONTEXT_KEY);
 
 	export let layer: Layer;
+	export let metadata: VectorTileMetadata;
 
-	const parentLayerId = layer.id;
+	let parentLayerId = layer.id;
 	let style: LayerSpecification = getLayerStyle($map, layer.id);
 	let textFieldValue = '';
-	let isAdvancedSettings = false;
 	let onlyNumberFields = false;
 	let targetLayer: Layer = style.type === 'symbol' ? layer : undefined;
 	let targetLayerId = targetLayer ? layer.id : `${parentLayerId}-label`;
+
+	let tabs = [
+		{
+			label: 'font',
+			icon: 'fa-solid fa-font'
+		},
+		{
+			label: 'color',
+			icon: 'fa-solid fa-fill'
+		}
+	];
+	let activeTab: string = tabs[0].label;
 
 	onMount(() => {
 		initialiseTextLabel();
@@ -57,96 +79,98 @@
 </script>
 
 {#if targetLayer}
-	<div class="action" data-testid="vector-label-panel-container">
-		<div class="columns is-mobile is-10 mb-0 is-vcentered is-justify-content-space-between">
-			<div class="column is-3">Property:&nbsp;</div>
-			<div class="column pl-0 pr-5 is-7">
+	<div class="label-container py-2">
+		{#if textFieldValue && $map.getLayer(layer.id)}
+			<div class="editor-button">
+				<VectorLabelPropertyEditor bind:layerId={targetLayer.id} bind:parentId={parentLayerId} />
+			</div>
+		{/if}
+
+		<FieldControl title="Property">
+			<div slot="help">Select a property to show data label for a vector layer.</div>
+			<div slot="control">
 				<TextField bind:onlyNumberFields on:change={fireLabelChanged} bind:layer={targetLayer} />
 			</div>
-		</div>
+		</FieldControl>
+
 		{#if textFieldValue && $map.getLayer(layer.id)}
 			{@const fieldType = getTextFieldDataType($map, layer, textFieldValue)}
-			{#if fieldType && ['number', 'float'].includes(fieldType)}
-				<div class="columns is-mobile is-12 m-auto is-vcentered">
-					<div class="column is-8 pl-0">Number of decimal places</div>
-					<div class="column is-3 is-flex is-justify-content-center">
-						<TextFieldDecimalPosition bind:layerId={targetLayer.id} />
-					</div>
-				</div>
-			{/if}
 
-			<div class="columns is-mobile is-12 mb-0 pb-0 is-vcentered">
-				<div class="column is-3 pr-0">Font color:</div>
-				<div class="column pl-0 is-1">
-					<TextColor bind:layerId={targetLayer.id} />
-				</div>
-				<div class="column is-3 pl-4 pr-0">Font size:</div>
-				<div class="column pl-0 is-5">
-					<TextSize bind:layerId={targetLayer.id} />
-				</div>
-			</div>
-			<div class="columns is-mobile is-12 mb-0 pb-0 is-vcentered">
-				<div class="column is-3 pr-0">Halo color:</div>
-				<div class="column pl-0 is-1">
-					<TextHaloCalor bind:layerId={targetLayer.id} />
-				</div>
-				<div class="column is-3 pl-4 pr-0">Halo width:</div>
-				<div class="column pl-0 is-5">
-					<TextHaloWidth bind:layerId={targetLayer.id} />
-				</div>
+			<div class="tabs is-centered is-toggle">
+				<ul>
+					{#each tabs as tab}
+						<li class={activeTab === tab.label ? 'is-active' : ''}>
+							<!-- svelte-ignore a11y-missing-attribute -->
+							<a
+								class="has-text-weight-bold"
+								role="tab"
+								tabindex="0"
+								data-sveltekit-preload-code="off"
+								data-sveltekit-preload-data="off"
+								on:click={() => {
+									activeTab = tab.label;
+								}}
+								on:keydown={handleEnterKey}
+							>
+								<span class="icon is-small"><i class={tab.icon} aria-hidden="true"></i></span>
+								<span class="is-capitalized">{tab.label}</span>
+							</a>
+						</li>
+					{/each}
+				</ul>
 			</div>
 
-			<div class="columns is-mobile advanced-settings">
-				<div class="column is-6 m-auto">
-					<div class="field">
-						<input
-							id="switchAdvancedSettings-{layer.id}"
-							type="checkbox"
-							name="switchSmall-{layer.id}"
-							class="switch is-small is-rounded is-info"
-							bind:checked={isAdvancedSettings}
-						/>
-						<label for="switchAdvancedSettings-{layer.id}" class="is-size-6"
-							>Advanced Settings</label
-						>
-					</div>
-				</div>
-			</div>
+			<div hidden={activeTab !== tabs[0].label}>
+				<FieldControl title="Font">
+					<div slot="help">The text font with which the text will be drawn.</div>
+					<div slot="control"><TextFont bind:layerId={targetLayer.id} /></div>
+				</FieldControl>
 
-			{#if isAdvancedSettings}
-				<div class="advanced-settings-container pb-4">
-					<div class="columns is-mobile">
-						{#if style.type === 'fill' || style.type === 'line'}
-							<div class="column">
-								<div class="has-text-centered pb-2">Label position relative to geometry</div>
-								<div class="is-flex is-justify-content-center">
-									<SymbolPlacement bind:layer={targetLayer} />
-								</div>
+				<div class="grid">
+					<FieldControl title="Font size">
+						<div slot="help">The font size with which the text will be drawn.</div>
+						<div slot="control"><TextSize bind:layerId={targetLayer.id} /></div>
+					</FieldControl>
+
+					{#if fieldType && ['number', 'float'].includes(fieldType)}
+						<FieldControl title="Decimal position">
+							<div slot="help">
+								The number of decimal places with which the numeric value label will be formated.
 							</div>
-						{/if}
-
-						<div class="column">
-							<div class="has-text-centered">Maximum width text wrap</div>
-							<div class="is-flex is-justify-content-center" style="position: relative;">
-								<TextMaxWidth bind:layerId={targetLayer.id} />
-							</div>
-						</div>
-					</div>
+							<div slot="control"><TextFieldDecimalPosition bind:layerId={targetLayer.id} /></div>
+						</FieldControl>
+					{/if}
 				</div>
-			{/if}
+			</div>
+
+			<div hidden={activeTab !== tabs[1].label}>
+				<TextColor
+					bind:layerId={targetLayer.id}
+					bind:metadata
+					classesContextKey={NUMBER_OF_CLASSES_CONTEXT_KEY_LABEL}
+					colorContextKey={DEFAULTCOLOR_CONTEXT_KEY_LABEL}
+					colormapContextKey={COLORMAP_NAME_CONTEXT_KEY_LABEL}
+				/>
+			</div>
 		{/if}
 	</div>
 {/if}
 
 <style lang="scss">
-	.advanced-settings-container {
-		padding-left: 15px;
-		padding-right: 15px;
+	.label-container {
+		position: relative;
 
-		> .columns {
-			border: 1px solid #ccc;
-			padding: 0;
-			padding-top: 5px;
+		.editor-button {
+			position: absolute;
+			top: 0em;
+			right: 0em;
+			z-index: 10;
+		}
+
+		.grid {
+			display: grid;
+			grid-template-columns: repeat(2, 1fr);
+			gap: 10px;
 		}
 	}
 </style>