+++ sites/geohub/src/stores/index.ts
@@ -7,3 +7,5 @@ export * from './colorMapName';
 export * from './classificationMethod';
 export * from './defaultColor';
 export * from './headerHeight';
+export * from './PageDataLoading';
+export * from './progressBar';